se parametrizo el valor de la compilacion

+++ escritorio/src/main/java/bo/firmadigital/jacobitus4/localhost9000/dtos/VersionDto.java
@@ -10,6 +10,6 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class VersionDto {
-    private Integer compilacion;
+    private Long compilacion;
     private String api_version;
 }
